server/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
pillow==10.4.0
safetensors==0.4.4
accelerate==0.33.0
transformers==4.44.2
diffusers==0.30.2
xformers==0.0.27.post2
